Restricted Pages

+++ studybud/base/views.py
@@ -4,7 +4,9 @@ from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
+from django.http import HttpResponse
 
 # rooms = [
 #     {'id' : 1,'name' : 'Lets learn python'},
@@ -28,10 +30,11 @@ def loginPage(request):
             login(request, user)
             return redirect('home')
         else:
-            messages.errors(request,'Username or Password doesnot exists')
+            messages.errors(request, 'Username or Password doesnot exists')
 
-    context = {'user':user}
+    context = {}
     return render(request, 'base/login_register.html', context)
+
 
 def logoutUser(request):
     logout(request)
@@ -58,6 +61,7 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='/login')
 def createRoom(request):  # -> tuple[Any, Literal['room_form.html'], dict]:
     form = RoomForm()
     if (request.method == 'POST'):
@@ -70,9 +74,13 @@ def createRoom(request):  # -> tuple[Any, Literal['room_form.html'], dict]:
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -83,8 +91,12 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
